fix(front): sort by creation date as secondary criterion
As soon as there would be multiple issues, new issues will be completely at the bottom, so it would get hard to find them.
Also more recent issues with the same amount of upvotes as older issues might be more important, bc they got to that number in less time.