refactor: unhack the column width fix
When clicking on a column to sort it, it would change the width of the column in some cases since when it is sorted an icon is shown.
The refactor now conditionally adds a margin on the right which takes the place of the icon when it is not displayed.