feat: remove scrollbar and resizing from input fields
Given that the input fields automatically resize with the amount of text inside them, the scrollbars are not really necessary anymore and looked outdated.

Removing the resizing option would only serve to make the input field smaller, which is not really useful.

Both of these changes make the input fields just a bit cleaner.