revert: add back issue ID to title on student server
I know we agreed that it's not necessary to have the ID in the title.
But I just figured that we haven't considered that that opens up possibilities for people to troll.
They can just create an issue with the same title as a hot issue but write some spam inside it.